+++ packages/@expo/metro-config/src/serializer.ts
@@ -10,6 +10,8 @@ import baseJSBundle from 'metro/src/DeltaBundler/Serializers/baseJSBundle';
 import bundleToString from 'metro/src/lib/bundleToString';
 import countLines from 'metro/src/lib/countLines';
 
+import { env } from './env';
+
 const debug = require('debug')('expo:metro-config:serializer') as typeof console.log;
 
 type Serializer = NonNullable<ConfigT['serializer']['customSerializer']>;
@@ -78,16 +80,28 @@ export function serializeWithEnvironmentVariables(
   if (options.dev) {
     // Set the process.env object to the current environment variables object
     // ensuring they aren't iterable, settable, or enumerable.
-    const str = `Object.defineProperty(process, 'env', {
-      value: Object.freeze(Object.defineProperties({}, {
-        ${Object.keys(getAllExpoPublicEnvVars())
-          .map((key) => `${JSON.stringify(key)}: { value: ${JSON.stringify(process.env[key])} }`)
-          .join(',')}
-      }))
-    });`;
+    const str = `process.env=Object.defineProperties(process.env, {${Object.keys(
+      getAllExpoPublicEnvVars()
+    )
+      .map((key) => `${JSON.stringify(key)}: { value: ${JSON.stringify(process.env[key])} }`)
+      .join(',')}});`;
 
-    const envCode = `var process=this.process||{};${str}`;
-    return [entryPoint, [getEnvPrelude(envCode), ...preModules], graph, options];
+    const [firstModule, ...restModules] = preModules;
+    // const envCode = `var process=this.process||{};${str}`;
+    // process.env
+    return [
+      entryPoint,
+      [
+        // First module defines the process.env object.
+        firstModule,
+        // Second module modifies the process.env object.
+        getEnvPrelude(str),
+        // Now we add the rest
+        ...restModules,
+      ],
+      graph,
+      options,
+    ];
   }
 
   // In production, inline all process.env variables to ensure they cannot be iterated and read arbitrarily.
@@ -129,7 +143,12 @@ function getEnvPrelude(contents: string): Module<MixedOutput> {
 }
 
 export function withExpoSerializers(config: InputConfigT): InputConfigT {
-  return withSerialProcessors(config, [serializeWithEnvironmentVariables]);
+  const processors: SerialProcessor[] = [];
+  if (!env.EXPO_NO_CLIENT_ENV_VARS) {
+    processors.push(serializeWithEnvironmentVariables);
+  }
+
+  return withSerialProcessors(config, processors);
 }
 
 // There can only be one custom serializer as the input doesn't match the output.